requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.49
alembic==1.12.0
psycopg2-binary==2.9.12
apscheduler==3.11.1
alpaca-py==0.43.4
ib_insync==0.9.86
anthropic==0.40.0
openai==1.55.0
google-genai==0.8.0
yfinance==0.2.35
pandas==2.2.2
python-dotenv==1.0.0
pydantic==2.13.4
websockets==13.0
pytz==2026.2
email-validator==2.0.0
